+++ drel/django/__init__.py
@@ -1,1 +1,1 @@
-from .api import LoggingMiddleware
+from .api import LoggingMiddleware, mail_admins_on_es_exception